test: test contributor grouping strategy
