+++ src/components/modals/GroupContributorsModal.tsx
@@ -17,6 +17,7 @@ import { useViewSubmit } from "~/hooks"
 import { Modal } from "~/components/modals/Modal"
 import { missingInMapColor } from "~/const"
 import { cn } from "~/styling"
+import { autoBuildContributorGroups } from "~/components/modals/utils/autoBuildContributorGroups"
 
 export function GroupContributorsModal({ open, onClose }: { open: boolean; onClose: () => void }) {
   const { databaseInfo } = useData()
@@ -27,8 +28,8 @@ export function GroupContributorsModal({ open, onClose }: { open: boolean; onClo
   const [, contributorColors] = useMetrics()
   const [, startTransition] = useTransition()
   const submit = useViewSubmit()
-  // When the modal is closed, check if the contributor groups have changed. If they have, submit the new groups as custom close action.
 
+  // When the modal is closed, check if the contributor groups have changed. If they have, submit the new groups as custom close action.
   const handleCloseWithSubmitIfChanged = () => {
     const currentSerialized = JSON.stringify(localContributorGroups)
     const originalSerialized = JSON.stringify(contributorGroups)
@@ -72,7 +73,7 @@ export function GroupContributorsModal({ open, onClose }: { open: boolean; onClo
   }, [ungroupedContributorsSorted, filter])
 
   const autoGroupCandidates = useMemo(
-    () => buildAutoContributorGroups(ungroupedContributorsSorted),
+    () => autoBuildContributorGroups(ungroupedContributorsSorted),
     [ungroupedContributorsSorted]
   )
   const autoGroupAuthorCount = useMemo(
@@ -367,60 +368,4 @@ function AliasEntry({
 
 function uniqueId(person: Person) {
   return `${person.name}\u0000${person.email ?? ""}`
-}
-
-function buildAutoContributorGroups(ungroupedContributors: Person[]): ContributorGroup[] {
-  if (ungroupedContributors.length < 2) return []
-
-  const stringSorter = (a: string, b: string) => a.toLowerCase().localeCompare(b.toLowerCase())
-  const parent = Array.from({ length: ungroupedContributors.length }, (_, i) => i)
-  const rank = Array(ungroupedContributors.length).fill(0)
-
-  const find = (i: number): number => {
-    if (parent[i] !== i) parent[i] = find(parent[i])
-    return parent[i]
-  }
-
-  const union = (a: number, b: number) => {
-    const rootA = find(a)
-    const rootB = find(b)
-    if (rootA === rootB) return
-    if (rank[rootA] < rank[rootB]) parent[rootA] = rootB
-    else if (rank[rootA] > rank[rootB]) parent[rootB] = rootA
-    else {
-      parent[rootB] = rootA
-      rank[rootA]++
-    }
-  }
-
-  const firstByName = new Map<string, number>()
-  const firstByEmail = new Map<string, number>()
-
-  for (const [i, { name, email }] of ungroupedContributors.entries()) {
-    if (name) {
-      if (firstByName.has(name)) union(i, firstByName.get(name)!)
-      else firstByName.set(name, i)
-    }
-    if (email) {
-      if (firstByEmail.has(email)) union(i, firstByEmail.get(email)!)
-      else firstByEmail.set(email, i)
-    }
-  }
-
-  const groupsByRoot = new Map<number, Person[]>()
-  for (const [i, person] of ungroupedContributors.entries()) {
-    const root = find(i)
-    const group = groupsByRoot.get(root) ?? []
-    group.push(person)
-    groupsByRoot.set(root, group)
-  }
-
-  return Array.from(groupsByRoot.values())
-    .filter((g) => g.length > 1)
-    .map((members) => {
-      const sorted = [...members].sort(
-        (a, b) => stringSorter(a.name, b.name) || stringSorter(a.email ?? "", b.email ?? "")
-      )
-      return { displayName: sorted[0]?.name ?? "Group", members: sorted }
-    })
 }
